security_camera/requirements_raspi.txt
# requirement file for raspberry
--extra-index-url https://google-coral.github.io/py-repo/
tflite-runtime==2.14.0
numpy==1.26.4
scipy==1.13.1
sounddevice==0.4.6
soundfile==0.12.1
opencv-python==4.9.0.80
pillow==10.3.0
